backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.12
SQLAlchemy==2.0.36
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
chromadb==0.6.3
insightface==0.7.3
onnxruntime==1.20.1
numpy==1.26.4
Pillow==11.1.0
pillow-heif==0.21.0
python-magic==0.4.27
python-dotenv==1.0.1
bcrypt==4.2.1
gunicorn==23.0.0
uuid6==2024.7.10
rank_bm25==0.2.2
scikit-learn>=1.3.0
